requirements.txt
python-nest
python-forecastio
